feat: add title component
