+++ src/components/app-layout/app-layout.tsx
@@ -4,8 +4,13 @@ import './css/app-layout.css';
 
 interface AppLayoutPropsI {
   children: ReactNode;
+  bgColor?: string;
 }
 
-export function AppLayout({ children }: AppLayoutPropsI) {
-  return <Box className='app__layout'>{children}</Box>;
+export function AppLayout({ children, bgColor }: AppLayoutPropsI) {
+  return (
+    <Box className='app__layout' bgcolor={bgColor}>
+      {children}
+    </Box>
+  );
 }
