+++ src/components/app-layout/css/app-layout.css
@@ -1,4 +1,5 @@
 .app__layout {
   margin: 0 auto;
-  width: 1000px;
+  padding: 0 24px;
+  width: 900px;
 }
