+++ src/components/country-app/country-app-title.tsx
@@ -1,0 +1,10 @@
+import { Typography } from '@mui/material';
+import './css/country-app.css';
+
+export function CountryAppTitle() {
+  return (
+    <Typography className='country__app-title' variant='h5' fontWeight='700'>
+      Countries
+    </Typography>
+  );
+}
