+++ src/components/country-app/css/country-app.css
@@ -1,0 +1,9 @@
+.country__app__box {
+  border-left: 1px solid black;
+  border-right: 1px solid black;
+}
+
+.country__app-title {
+  color: #455a64;
+  font-weight: 700;
+}
