+++ src/components/header/css/header.css
@@ -3,3 +3,7 @@
   justify-content: center;
   height: 75px; 
 }
+
+.header__toolbar {
+ padding: 0 !important;
+}
