+++ src/index.tsx
@@ -1,12 +1,28 @@
+import { CssBaseline, ThemeProvider, createTheme } from '@mui/material';
+import { blueGrey } from '@mui/material/colors';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 
+const theme = createTheme({
+  palette: {
+    primary: {
+      light: blueGrey[300],
+      main: blueGrey[500],
+      dark: blueGrey[700]
+    }
+  }
+});
+
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
+
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <ThemeProvider theme={theme}>
+    <CssBaseline />
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </ThemeProvider>
 );
